chore: rename ports to infrastructure_contracts
