+++ src/libecalc/domain/energy/infrastructure_contracts.py
